Create Number Pattern
Write a program that reads an input n from stdin. n tells us the number of rows to be printed. Print the pattern as per the given examples.

Input format: The first line contains the number of inputs T. The lines after that contain a different values for n

Output format: Print the pattern as per the example

Example Input:
2
3
5
Output:
1
1 2
1 2 3
1
1 2
1 2 3
1 2 3 4
1 2 3 4 5